Update bluez.py read_local_bdaddr()
When we use the BD_ADDR as a string, each byte of the BD_ADDR should be converted into two hexadecimal digits, not one hexadecimal digit. Otherwise we might fail to call hci_devid() using the return value, for example ['0:XX:XX:XX:XX:XX'], of the read_local_bdaddr().